+++ docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/docs
 
+## 2.7.22
+
+### Patch Changes
+
+- Various changes
+
+## 2.7.21
+
+### Patch Changes
+
+- Various changes
+
+## 2.7.20
+
+## 2.7.19
+
 ## 2.7.18
 
 ### Patch Changes