+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @stackframe/js-example
 
+## 2.7.22
+
+### Patch Changes
+
+- @stackframe/js@2.7.22
+
+## 2.7.21
+
+### Patch Changes
+
+- @stackframe/js@2.7.21
+
+## 2.7.20
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/js@2.7.20
+
+## 2.7.19
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/js@2.7.19
+
 ## 2.7.18
 
 ### Patch Changes