+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @stackframe/stack-emails
 
+## 2.7.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.22
+  - @stackframe/stack-ui@2.7.22
+
+## 2.7.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.21
+  - @stackframe/stack-ui@2.7.21
+
+## 2.7.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.7.20
+  - @stackframe/stack-ui@2.7.20
+
+## 2.7.19
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.7.19
+- @stackframe/stack-ui@2.7.19
+
 ## 2.7.18
 
 ### Patch Changes